farmpolicyによる認可の実装

+++ app/Http/Controllers/FarmsController.php
@@ -5,18 +5,22 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\User;
 use App\Farm;
+use App\FarmUser;
 use Auth;
+use Gate;
 
 use Illuminate\Support\Facades\DB;
 use App\Http\Requests\StoreFarmData;
 
 class FarmsController extends Controller
 {
+    
+    
     public function index()
     {
-       $farmDatas = Farm::orderBy('id', 'desc')->get();
-       
-       return view('farms.index', compact('farmDatas'));
+        
+        $farmDatas = Farm::orderBy('id', 'desc')->get();
+        return view('farms.index', compact('farmDatas'));
       
     }
 
@@ -29,7 +33,7 @@ class FarmsController extends Controller
     
     public function store(StoreFarmData $request, Farm $farm)
     {
-        //fill関数に変更
+         //fill関数に変更
          $farm->fill($request->all())->save();
          //中間テーブルへレコードを追加。
          $farm->users()->sync(Auth::user()->id);
@@ -39,22 +43,28 @@ class FarmsController extends Controller
  
     public function show($id)
     {
-       $farmData = Farm::find($id);
+        $farmData = Farm::find($id);
 
-       return view('farms.show', compact('farmData'));
+        return view('farms.show', compact('farmData', 'farm'));
     }
 
     
     public function edit($id)
     {
-       $farmData = Farm::find($id);
-       return view('farms.edit', compact('farmData'));
+        //policyによる認可処理
+        $farmData = Farm::find($id);
+        Gate::authorize('update', $farmData);
+        
+        return view('farms.edit', compact('farmData'));
     }
 
 
     public function update(StoreFarmData $request, $id)
     {
+        //policyによる認可処理
         $farmData = Farm::find($id);
+        Gate::authorize('update', $farmData);
+
         $farmData->fill($request->all())->save();
 
         return view('farms.show', compact('farmData'));
@@ -68,7 +78,10 @@ class FarmsController extends Controller
      */
     public function destroy($id)
     {
-        $farmData = Auth::user()->farms()->find($id);
+        //policyによる認可処理
+        $farmData = Farm::find($id);
+        Gate::authorize('delete', $farmData);
+        
         $farmData->delete();
         return view('farms.show', compact('farmData'));
     }
